+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file contains relevant information for each release. Please refer to the commit history for more details.
 
+### [0.36.0](https://github.com/davidledwards/ped/tree/v0.36.0) | `2025-11-30`
+
+#### Added
+
+- CLI option `--syntaxes` to print available syntaxes
+- Key `M-h s` bound to new `help-syntaxes` operation, which opens `@syntaxes` window
+
 ### [0.35.0](https://github.com/davidledwards/ped/tree/v0.35.0) | `2025-11-09`
 
 #### Added